fix key repeat by moving unsigned int time to double time

+++ game/shared.h
@@ -73,9 +73,9 @@ typedef struct {
 
 typedef struct {
 	int			down[2];		// key nums holding it down
-	unsigned	firstdowntime;
-	unsigned	downtime;		// msec timestamp
-	unsigned	msec;			// msec down this frame if both a down and up happened
+	double	firstdowntime;
+	double	downtime;		// msec timestamp
+	double	msec;			// msec down this frame if both a down and up happened
 	bool		active;			// current state
 	bool		wasPressed;		// set when down, not cleared when up
 } kbutton_t;
